fix(test): Fix test about default storage type for DefaultStorageMediaProviderTest (#2419)
### What changes were proposed in this pull request?
This part of testStorageProvider should test the default storage media for a non-existent directory. But 47cac39 changed the behavior of getStorageMediaFor, now it checks the parent directory (recursively) so for /path/to/base/dir it will end up checking / which is existent. With this change it will check path which is missing, so it will really check the default value.

### Why are the changes needed?

### Does this PR introduce any user-facing change?
No.

### How was this patch tested?
UT