[#2457] feat(spark): Introducing shuffle-server data push statistics (#2458)
### What changes were proposed in this pull request?

Introducing shuffle-server data push statistics in writer

### Why are the changes needed?

Sometimes we found some jobs are slow due to the underlying one server slow, but under the current codebase, it's hard to found it. This PR is to track the shuffle-servers speed to output.

BTW, based on this PR, I will introduce the uniffle spark UI for easier and better observability.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Internal tests.