[#2583] fix(spark): Enable taskIds filter only on AQE and multi replicas for reader (#2584)
### What changes were proposed in this pull request?

Only enable taskIds filter mechanism on AQE skew hit and multi replicas. Previously, when multi shuffle servers are assigned for reader to read, it will enable this mechanism, that will hurt the shuffle-servers performance due to the bitmap check, actually there is no need to do this on partition reassign (partition split) is enabled.

### Why are the changes needed?

To fix #2583 

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Existing tests