[#2601] feat(spark): Introduce overlapping decompression for shuffle read (#2602)
### What changes were proposed in this pull request?

This PR is to introduce the overlapping decompression for the shuffle reading for the better shuffle speed.

### Why are the changes needed?

for #2601

When applying the https://github.com/apache/uniffle/pull/2598 into the benchmark of terasort 100g, I found some bottleneck for the decompression time in the read phase. 

Based on a 100 GB Terasort benchmark, the results are impressive, reducing shuffle read time by 50% after applying this PR.

<img width="1911" height="716" alt="image" src="https://github.com/user-attachments/assets/707598fb-0850-4d38-8453-f34f852fc6af" />


### Does this PR introduce _any_ user-facing change?

Yes

### How was this patch tested?

1. Unit tests.